Rename Type to FieldType
In the future, I'm going to try to add a Field class to clean up some of
these rough spots. For now, just renaming Type to FieldType and type.py
to field.py so that imports will work better. Also moved over the rest
of the imports in non-test to be of the approved style