Refactor model.py to clean it up
Thie file basically has three different types of methods: methods that
help define a model schema and interact with it (now in the ModelBase
class), methods that use the model schema to interact with Spanner (now
in the ModelMeta class), and objects that represent rows of the table
associated with the model schema (now in the Model class)

As part of this, I got rid of a bunch of @classmethods. Now methods from
the first two categorie have to be called on a model class and methods
from the third category have to be called on a model object instance.
This means that some of the methods in the Model class have to call
methods on the metaclass, which might be a bit odd.

Also some of the methods have turned into properties. This was aided by
getting rid of the @classmethods, as a method can't be both a
@classmethod and a @property