Add OrCondition
Some things have been added in a related project such that I want to
query against a list of things, except the things now need two columns
to be specified properly. The way I'm planning to solve this is do an OR
across multiple of these pairs, which requires support for ORs in the
ORM.
While I was in here, I cleaned up some of the other condition
implementations.