Remove values dictionary from Model instances
Starting work on improving performance of spanner_orm. Keeping around
separate dictionaries for field values and related items is slower than
storing it in __dict__, and also there's no real reason to do it. To not
break existing clients using .values, I added a `values` property that
constructs the dictionary from all fields.
One changed test is about changing a list after assigning it to a model,
for which I'm not sure the original behavior was actually desired. The
other is that now None values show up in the `values` property, which
shouldn't break any existing behavior