Skip validation on values from Spanner
If the `persisted` flag is True during model instance creation, we
should assume that the values came from Spanner, and thus we don't have
to validate them. Otherwise, we should validate all field values.

Originally we hadn't been running validation on all field values, so we
were allowing some fields to be None when they were not set as nullable.
I've updated the tests so that they pass again after this change.