Skip calling __setattr__ during model __init__
I'm a little less certain about this change: it makes model
initialization significantly faster, but it does so by directly using
__dict__ (so we skip hitting our overridden __setattr__ method). This
makes it incompatible with using __slots__ later on. On the other hand,
I wasn't super happy with calling super().__setattr__() previously.
I also updated __setattr__ to make it a bit faster when setting
non-model attributes (`in` on a list is somewhat slow, so we avoid
checking primary keys until we have to)