Add find_multi
Adding a method that will make it easier to find multiple rows from
Spanner at the same time, because using the find API is faster than
using execute_sql with a WHERE IN UNNEST(). I debated changing the
find method so that it allows finding multiple rows, but couldn't
decide one way or the other. Thoughts?