Add code to run migrations against a given Spanner database
After migrations are written, they need to be run against a given
database to be applied. I've added a model to store the migration status
of migrations on the current database and then wrote the executor to
figure out which migrations need to be executed (both for upgrading and
downgrading migrations), execute them, and then update the database to
indicate that the migration has been run.
In talking with @dseomn, I realized that 'project' can be inferred by
the Spanner client libraries, so I've moved that to be an optional
parameter.