Reorganize the code related to updating database ddl
Moved ddl update code and the schema models into admin, as these are
only used in admin-related functionality. Took the SchemaUpdate
execution code out of spanner_orm.admin.metadata (as it's not really
related to retrieving metadata, apart from it using that functionality)
and put it into SchemaUpdate, where it belongs. Also updated
SchemaUpdate to implement more of the missing functionality