Add reading and writing of migration files
Migration files are going to be pretty simple, they contain an id, the
id of the migration the current one depends on, and methods to apply and
rollback the migration that each return a single SchemaUpdate object.
I've added the code to write a migration file, read all migration files,
and order the migrations based on the 'prev_migration' field to
determine what order the migrations have to be applied in.