Remove creation_ddl from model (#42)
This really belongs in the update code, so move it from model.Model to
admin.update.CreateTable. Also because creating the table fresh on a new
database also requires involve adding secondary indices, add a helper
method that takes a model and returns the list of ddl necessary to set
up the table for that model.