Give all ModelMetaclass classes a `meta` attribute.
This fixes bugs where pytype got the type for `meta` wrong because it
only existed on some classes. I'm not 100% sure adding a default (empty)
metadata to classes that didn't previously have it makes sense, but
tests pass and it seems likely to be safe to me.

This change also has the side effect of parsing Model-specific
attributes from the Model class itself (as opposed to just subclasses of
Model), but it doesn't look like Model has any attributes that would be
affected.

Addresses #93.

Tested: Ran tests and pytype.