Make option-like parameters (mostly `transaction`) keyword-only.
Note that this change breaks the API, but I think it's worth it. In some
cases, e.g., where(), this avoids needing to pass an unnamed `None`
argument which I think is confusing. In the case of save_batch()'s
`force_write` parameter, I think this can avoid potentially bad bugs.
And in all cases, I think this makes the call sites more clear.

Additionally:

1. Make the `transaction` parameter optional for all methods.
2. Fix some type annotations on those parameters.