Make Param type inference explicit.
This makes it possible to use a more restrictive type annotation than
Any, with the main benefit of excluding Optional[primitive] types. With
the old Any, I think there was a risk of code passing
Optional[primitive], only testing when the value isn't None, and getting
errors in production when the value is None. With this change, the type
checker should complain about that code.