Make it possible to add non-DDL migration updates.
SchemaUpdate seems to be designed for DDL (schema) changes, but there
currently doesn't seem to be any way to add a `NOT NULL` column using
only DDL changes. This commit makes it easier to add non-DDL migration
updates; a follow-up commit will add support for backfilling data in a
migration update.