Update for rudimentary base64 encoding check in BYTES type.
This performs a validation to allow for a base64 encoding check under the newly supported Bytes type. This allows for spanner_orm to be initially work with base64 encoded bytes and will be backward compatible with any future potential changes involving base64 conversions to support arbitrary bye encoding in the future.