Use instances of FieldType instead of classes
Aside from
<https://google.github.io/styleguide/pyguide.html#217-function-and-method-decorators>,
this makes it possible to have types with parameters. E.g., this makes
it possible to add code for something like Array(String(length=20)) for
the type ARRAY<STRING(20)>. That way we don't need to define a new
FooArray class for each type of array, and we can add support for
lengths other than MAX more easily.

This could break any users that use FieldType for anything other than
passing to Field, but hopefully the vast majority of uses just pass it
to Field.